requirements.txt
faiss-cpu==1.7.4
numpy==1.26.4
openai==1.14.3
PyPDF2==3.0.1
flask==3.1.0
flask-cors==4.0.0
sentence-transformers==2.3.1
transformers==4.33.2
huggingface_hub==0.16.4
werkzeug>=3.1.0
tf-keras
